requirements.txt
google-cloud-pubsub==2.27.2
pymongo==4.10.1
google-cloud-bigquery==3.29.0
python-dotenv==1.0.1